[#6241] YSQL: Make row key share lock less restrictive
Summary:
Two major scenarios are fixed by this diff:
1. FK check uses request with batch of ybctid. This request has no `partition_column_values` and due to error `PgsqlReadOperation::GetIntents` method takes a lock on a whole tablet. As a result the following scenario fails

```
CREATE TABLE parent(k int PRIMARY KEY) SPLIT INTO 1 TABLETS;
CREATE TABLE child(pk int PRIMARY KEY, CONSTRAINT parent_fk FOREIGN KEY(pk) REFERENCES parent(k)) SPLIT INTO 1 TABLETS;

-- Connection 1
SET yb_transaction_priority_lower_bound = 0.5;
BEGIN;
INSERT INTO parent VALUES(1);
INSERT INTO child VALUES(1);

-- Connection 2:
SET yb_transaction_priority_upper_bound = 0.1;
INSERT INTO parent VALUES(2);

ERROR:  Operation failed. Try again.: 7d53ad8f-bb92-4020-bd98-2ee07b676a3f Conflicts with higher priority transaction: 60582ba4-2623-4e8e-b1ba-960351e4bcfe
```

2. `DocDB` layer uses range key component while adding intents for row locking.
But YSQL requests only come with hash (partition) components even when the SQL query sets all range components .

Example:
```
CREATE TABLE t(h INT, r INT, PRIMARY KEY(h, r ASC));
SELECT * FROM t WHERE h = 1 AND r = 10 FOR KEY SHARE OF t;
```

Instead of specifying explicit key value with hash and range components YSQL send request with hash components only and `condition_expr` field for filtering.
```
partition_column_values {
  value {
    int32_value: 1
  }
}
...
condition_expr {
  condition {
    op: QL_OP_AND
    operands {
      condition {
        op: QL_OP_EQUAL
        operands {
          column_id: 11
        }
        operands {
          value {
            int32_value: 10
          }
        }
      }
    }
  }
}
```

Solution is to always set range components in case SQL query has it. And move it into `condition_expr` field of request in case it is necessary before sending the query. Because it is possible that not all range components are set.

```
CREATE TABLE t(h INT, r1 INT, r2 INT, r3 INT, PRIMARY KEY(h, r1 ASC, r2 ASC, r3 ASC));
SELECT * FROM t WHERE h = 1 AND r1 = 1 AND r3 = 3;
```

In this case `r2` component is not set as a result `r3` can't be bind as range column value. It will be placed into `condition_expr` field.

```
partition_column_values {
  value {
    int32_value: 1
  }
}
...
range_column_values {
  value {
    int32_value: 1
  }
}
condition_expr {
  condition {
    op: QL_OP_AND
    operands {
      condition {
        op: QL_OP_EQUAL
        operands {
          column_id: 13
        }
        operands {
          value {
            int32_value: 3
          }
        }
      }
    }
  }
}
```

In case `r2` field is specified

```
SELECT * FROM t WHERE h = 1 AND r1 = 1 AND r2 = 2 AND r3 = 3;
```
 `condition_expr` will be empty and all range column values will be specified:
```
partition_column_values {
  value {
    int32_value: 1
  }
}
...
range_column_values {
  value {
    int32_value: 1
  }
}
range_column_values {
  value {
    int32_value: 2
  }
}
range_column_values {
  value {
    int32_value: 3
  }
}
```

Test Plan:
New test cases were introduced

```
./yb_build.sh --java-test org.yb.pgsql.TestPgExplicitLocks
./yb_build.sh --java-test org.yb.pgsql.TestPgPushdownEfficiency
./yb_build.sh --java-test org.yb.pgsql.TestPgForeignKey
```

Reviewers: neil, alex, zyu, mihnea

Reviewed By: alex, zyu, mihnea

Subscribers: rsami, bogdan, kannan, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D9839

+++ src/yb/common/schema.h
@@ -611,7 +611,7 @@ class Schema {
   // caught. If an invalid schema is passed to this constructor, an
   // assertion will be fired!
   Schema(const vector<ColumnSchema>& cols,
-         int key_columns,
+         size_t key_columns,
          const TableProperties& table_properties = TableProperties(),
          const Uuid& cotable_id = Uuid(boost::uuids::nil_uuid()),
          const PgTableOid pgtable_id = 0)
@@ -632,7 +632,7 @@ class Schema {
   // assertion will be fired!
   Schema(const vector<ColumnSchema>& cols,
          const vector<ColumnId>& ids,
-         int key_columns,
+         size_t key_columns,
          const TableProperties& table_properties = TableProperties(),
          const Uuid& cotable_id = Uuid(boost::uuids::nil_uuid()),
          const PgTableOid pgtable_id = 0)
@@ -648,7 +648,7 @@ class Schema {
   // Reset this Schema object to the given schema.
   // If this fails, the Schema object is left in an inconsistent
   // state and may not be used.
-  CHECKED_STATUS Reset(const vector<ColumnSchema>& cols, int key_columns,
+  CHECKED_STATUS Reset(const vector<ColumnSchema>& cols, size_t key_columns,
                        const TableProperties& table_properties = TableProperties(),
                        const Uuid& cotable_id = Uuid(boost::uuids::nil_uuid()),
                        const PgTableOid pgtable_id = 0) {
@@ -661,7 +661,7 @@ class Schema {
   // state and may not be used.
   CHECKED_STATUS Reset(const vector<ColumnSchema>& cols,
                        const vector<ColumnId>& ids,
-                       int key_columns,
+                       size_t key_columns,
                        const TableProperties& table_properties = TableProperties(),
                        const Uuid& cotable_id = Uuid(boost::uuids::nil_uuid()),
                        const PgTableOid pgtable_id = 0);
